Create a Packet class (#2)
Create a Packet class to hold packets exchanged with the robot. This
class is more aware of packet members like device, command, id, payload,
and crc than a byte array. This should make actions like creating a
packet, parsing a packet, generating a crc for a packet, and checking
the crc of a packet more self contained.

Some of the other SDK structures like the pending responses array are
refactored to store and compare tuples of packet device, command, and id
values instead of an array slice of bytes. Python list comprehension is
used to search and filter the pending responses array.

TODO:
- add a response requested flag to the Packet type instead of packing a
separate boolean in the pending responses tuple
- have pending responses be a list of packets instead of tuple
- or use named tuple for pending responses
- use enums for device numbers, led animations, marker positions, etc